refactor(sql): split eval.rs into focused modules
